requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pick>=2.2.0
PyPDF2>=3.0.0
pdf2image>=1.16.3
pytesseract>=0.3.10
python-slugify>=8.0.1
